update(pattern_matching): mv opendir,readdir,closedir to newfile
